+++ lib/mc_markdown.rb
@@ -1,0 +1,11 @@
+require 'redcarpet'
+
+# Parsing
+require_relative 'mc_markdown/parsers'
+
+# Formatters
+require_relative 'mc_markdown/formatters'
+
+# Util
+require_relative 'mc_markdown/extensions'
+require_relative 'mc_markdown/renderers'