+++ lib/mc_markdown/parsers.rb
@@ -1,0 +1,3 @@
+require_relative 'parsers/block_tag'
+require_relative 'parsers/formatter'
+require_relative 'parsers/short_tag'